server/requirements.txt
alembic==0.9.5
bcrypt==3.1.3
cffi==1.10.0
click==6.7
dnspython==1.15.0
email-validator==1.0.2
factory-boy==2.9.2
Faker==0.8.3
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.3.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.13.6
marshmallow-sqlalchemy==0.13.1
pycparser==2.18
PyJWT==1.5.3
PyMySQL==0.7.11
python-dateutil==2.6.1
python-dotenv==0.6.5
python-editor==1.0.3
six==1.10.0
SQLAlchemy==1.1.14
Werkzeug==0.12.2
WTForms==2.1
